+++ internal/api/web/_app/immutable/entry/start.Ckg6t3S9.js
@@ -1,0 +1,1 @@
+import{o as e,t}from"../chunks/Cxkt1zF_.js";export{e as load_css,t as start};